Refactor: allPlanDates 함수를 getDatesArrFromStartEnd로 변경"
- 변경한 함수를 util 함수로 변경하여 import로 사용
- ChangePlanState 컴포넌트만 추가
- PlanTopBar 위치를 /common에서 /plan으로 폴더 구조 변경
