+++ src/components/plan/changePlanState/ChangePlanState.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+import React from 'react';
+
+const ChangePlanState = () => {
+  return (
+    <div>
+      <div>ChangePlanState</div>
+    </div>
+  );
+};
+
+export default ChangePlanState;
